changed int int16_t in hubo.h to work with both 32 and 64 bit, if this does not work then it might be in the time struct

+++ include/hubo.h
@@ -228,12 +228,12 @@ struct hubo_state {
     struct hubo_joint_status status[HUBO_JOINT_COUNT];
 	struct hubo_jmc_state driver[HUBO_JMC_COUNT];
     double time;
-	int refWait;
+	int16_t refWait;
 }hubo_state_t;
 
 struct hubo_ref {
 	double ref[HUBO_JOINT_COUNT];	///< joint reference
-	int mode[HUBO_JOINT_COUNT]; 	///< mode 0 = filter mode, 1 = direct reference mode
+	int16_t mode[HUBO_JOINT_COUNT]; 	///< mode 0 = filter mode, 1 = direct reference mode
 }hubo_ref_t;
 
 struct jmcDriver{
@@ -245,7 +245,7 @@ struct hubo_board_cmd {
 	hubo_d_cmd_t type;		// Type of command. This value is REQUIRED. 
 					// Enumerated in hubo-daemonID.h
 
-	int joint;			// Target joint. If the message is meant for an entire board,
+	int16_t joint;			// Target joint. If the message is meant for an entire board,
 					// then fill this value with any joint number belonging to that
 					// board. This value is REQUIRED (with a few exceptions).
 
@@ -253,7 +253,7 @@ struct hubo_board_cmd {
 					// Note: This might or might not be used depending on the 
 					// type of message. TODO: Figure out if 8 is sufficient (or excessive)
 
-	int iValues[8];			// Integer values for the message. This may or may not be used
+	int16_t iValues[8];			// Integer values for the message. This may or may not be used
 					// depending on the type of message. TODO: Figure out of 10 is sufficient
 	
 	double dValues[8];		// Double values for the message. This may or may not be used
@@ -261,7 +261,7 @@ struct hubo_board_cmd {
 };
 
 
-extern int hubo_debug;
+extern int16_t hubo_debug;
 
 #endif //HUBO_PRIMARY_H
